feat: generic type support for attachable
The attachable implementation uses String as key type.
If someone selects the same string in the top layer,
the value will be shaded in the lower layer.

Although we can isolate using the namespace of a string,
it is not an absolute security guarantee. Adding generics
does not impose a burden on using a string as the key type,
but it should be noted that the current default implementation
uses a hash map for storage, which requires the key to meet
the requirements of the hash map for the key type.

Resolves: #4